# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: The Ministry of Herbology API
  version: 1.10.0
  description: |
    Contract for the Ministry of Herbology. Frozen — only the maintainers
    changes this file, via a recorded decision. Every path declares the parts of the project that
    implements it with `x-parts of the project`; `tests/contract/` enforces that the
    running API matches this document.

    Units: millimetres, °C, litres, lux. Timestamps are RFC 3339 UTC.
servers:
  - url: /api/v1
...
paths:
...
  /grounds/layers:
    get:
      tags: [grounds]
      operationId: listMapLayers
      responses:
        '200':
          description: Floor plans and property surveys
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MapLayer' }
    post:
      tags: [grounds]
      summary: Upload a floor plan or survey image
      operationId: createMapLayer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:    { type: string, format: binary }
                name:    { type: string }
                kind:    { type: string, enum: [floor_plan, survey] }
                site_id: { type: string, format: uuid }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MapLayer' }
  /grounds/layers/{layer_id}/calibration:
    parameters:
      - { name: layer_id, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      tags: [grounds]
      summary: Set the pixel↔world calibration points (survey) or px→mm scale (plan)
      operationId: setCalibration
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Calibration' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MapLayer' }
  /grounds/layers/{layer_id}/image:
    parameters:
      - { name: layer_id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [grounds]
      summary: The layer's image bytes
      operationId: getMapLayerImage
      description: >-
        What `MapLayer.image_url` points at. Declared in 1.5.0
        because the map fetches it for **every** layer, so it is not an
        operational endpoint with no client contract — it was briefly exempted
        as one, and the maps module said plainly in #46 that the exemption was
        wrong rather than leaving it to be found later.

        The bytes are whatever the operator uploaded, served back unchanged, so
        the media type is the one the upload's own magic bytes identified — never
        the `Content-Type` the client claimed.
      responses:
        '200':
          description: The image
          content:
            image/png:  { schema: { type: string, format: binary } }
            image/jpeg: { schema: { type: string, format: binary } }
        '404':
          description: >-
            No such layer, or its file is missing from the volume — which is
            what a restore that took the database and not the image volume looks
            like.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '503':
          description: >-
            No image directory is configured, so this deployment cannot serve
            layer images at all. Names the variable.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /grounds/pins:
    get:
      tags: [grounds]
      summary: Every specimen pin on a layer
      operationId: listPins
      parameters:
        - { name: layer_id, in: query, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Pins
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Pin' }
    put:
      tags: [grounds]
      summary: Place or move a specimen's pin
      operationId: setPin
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Pin' }
      responses:
        '200':
          description: Placed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Pin' }

  # ------------------------------------------------------------------ journal
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: moh_session
    # the stack's own services, GET only. Nginx deletes this header
    # from every inbound request, so it is unusable from outside the stack.
    service:
      type: apiKey
      in: header
      name: X-Moh-Service-Token
    # the Lunette's device token, valid on /hub/lunette only.
    lunette:
      type: http
      scheme: bearer
...
  schemas:
    Error:
      type: object
      required: [detail]
      properties:
        detail: { type: string }
        code:   { type: string }
...
    SpecimenBrief:
      type: object
      required: [id, display_name]
      properties:
        id:           { type: string, format: uuid }
        display_name: { type: string }
        is_outdoor:   { type: boolean }
        thumb_url:    { type: [string, 'null'] }
...
    PixelPoint:
      type: [object, 'null']
      properties:
        x: { type: number }
        y: { type: number }
    MapLayer:
      type: object
      required: [id, name, kind, image_url, image_width_px, image_height_px]
      properties:
        id:              { type: string, format: uuid }
        site_id:         { type: string, format: uuid }
        name:            { type: string }
        kind:            { type: string, enum: [floor_plan, survey] }
        image_url:       { type: string }
        image_width_px:  { type: integer }
        image_height_px: { type: integer }
        scale_mm_per_px: { type: [number, 'null'] }
        calibration:     { $ref: '#/components/schemas/Calibration' }
        ordinal:         { type: integer }
    Calibration:
      type: object
      description: |
        Floor plans use CRS.Simple and need only `scale_mm_per_px`. Surveys need
        at least two pixel↔world point pairs to be georeferenced.

        **Which half applies depends on the layer's `kind`, and the request body
        does not carry it** — so the rule can only live in the API, and it does
        (422, with the reason). Stated here because the document otherwise
        permits PUTing survey points to a floor plan, which is nonsense a reader
        should not have to discover from a rejection:

        * `kind: floor_plan` — send `scale_mm_per_px` alone, **no `points`**. A
          plan is a flat drawing; it has a scale, not a place on Earth.
        * `kind: survey` — send **two or more** `points`. `scale_mm_per_px` is
          *derived* from them and returned; sending it is not how a survey is
          georeferenced.

        `scale_mm_per_px` is therefore meaningful on both kinds — it is not
        plan-only — which is why this stays one schema rather than two.

        **There is no separate "uncalibrated" spelling.** Nothing here is
        required, so `{"scale_mm_per_px": null, "points": []}` *is* the
        uncalibrated state, and that is what the API serves from upload until
        somebody calibrates. Whether a layer is calibrated is **derived** — a
        plan when `scale_mm_per_px > 0`, a survey when it has two or more points
        — never read off a null. the design explains why `MapLayer.calibration`
        was deliberately *not* widened to accept `null`.
      properties:
        scale_mm_per_px: { type: [number, 'null'] }
        points:
          type: array
          items:
            type: object
            required: [px, world]
            properties:
              px:    { type: array, items: { type: number }, minItems: 2, maxItems: 2 }
              world: { type: array, items: { type: number }, minItems: 2, maxItems: 2 }
    Pin:
      type: object
      required: [specimen_id, layer_id, px]
      properties:
        specimen_id: { type: string, format: uuid }
        layer_id:    { type: string, format: uuid }
        px:          { $ref: '#/components/schemas/PixelPoint' }
        specimen:    { $ref: '#/components/schemas/SpecimenBrief' }
...
security:
  - session: []